requirements.txt
fastapi
uvicorn
aiogram
sqlalchemy
python-dotenv
httpx
beautifulsoup4
lxml
